# custom-gpt-openapi-3.1.yaml
openapi: 3.1.0
info:
  title: Bet Bot Complete Knowledge API
  description: AI-powered sports betting analytics with real-time data access and machine learning predictions
  version: 1.0.0
servers:
  - url: https://bet-bot-blipton03.replit.app
    description: Production Bet Bot API
paths:
...
  /api/gpt/predict-team:
    post:
      operationId: predict_game_outcome  
      summary: Get AI prediction for team matchup
      description: Generate AI prediction for any MLB team matchup with confidence scores and betting recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                homeTeam:
                  type: string
                  description: Home team name
                awayTeam:
                  type: string
                  description: Away team name
              required:
                - homeTeam
                - awayTeam

      responses:
        '200':
          description: Today's games with AI predictions
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                  totalGames:
                    type: integer
                  games:
                    type: array
                    items:
                      type: object
                  lastUpdated:
                    type: string
                  apiStatus:
                    type: string
...